# ARROW-9452: [Rust] [DataFusion] Optimize ParquetScanExec