@andygrove 2020-07-14
If nobody objects, I'd quite like this to make it in for the 1.0.0 release.